requirements.txt
Django
Pillow
django-import-export